requirements.txt
python-documentcloud==3.0.6
requests
